Backup ParticleController
Progress check on the implementation of ParticleController. Declared most things as if it was going to be a custom type, switched it to be a general wrapper class instead. Only MainModule is defined so far with the declaration for everything else commented out. Testing things on the global dust, making it larger, faster, and different colors.