Rename repository to Flair and clean project
Renamed the repository and all files to use Flair instead of Flare. Also cleaned a lot of the junk files from testing different things.
Flare was the second best name idea and a typo that I missed in earlier stages of development. Flair wins because I like how it emphasizes the user's creativity. Clutter removed because after a month of not having time, I've forgotten what all of it was for lol.