Use C# types in the Window custom types
Wrapped all of the pointers inside Window, and moved as many fields and methods as possible to C# types.
Wrapping the variables provides for an easier way to check both variable nullptr and object null. Declaring things as C# allows il2cpp and QRUE to see them, and if a few things such as Update() need to be declared as C# then I want everything* to be C#. No mixing and mashing C# and C++ for no good reason.