Add text next to FloatSettings and MinMaxCurveSettings
Added the setting's name next to the input instead of being the placeholder. New placeholder is 0.0.
Showing the name to the left makes it always visible and quick to read, as opposed to before where you had to delete the current text or hover over it for the hover hint. The new placeholder text of 0.0 shows that you can write a decimal and if nothing is written then the value defaults to 0.