Refactor project.cpp and established consistent naming
Renamed project.cpp to Project/import.cpp. Moved debug logHierarchy(...) to logging.cpp. Replaced int depth parameters with depthStr. Functions now use camelCase instead of PascalCase.
Splitting the Project class into multiple files will help make things more readable and easier to navigate. The debug logging function could be permanent, but it should be moved to a more general logging file. The component setup functions needed logging but it seemed overkill for them to calculate the string themselves; every function was changed to recursively concatenate the depth string. I still have no clue which naming system is "best," but I want to reason that camelCase makes sense in C++ land (a function is essentially a regular pointer that happens to be pointing to executable memory) even if my eyes are not fully used to seeing it.