Add path names to Project class
Each type of asset in the Project class is now a std::pair containing a string and the data. The string holds a pseudo file path, called pathName, that will help construct the file browser.
If a project has a lot of files, organization will be nice. They aren't functional at the moment but changing the data type now will save a future refactor.