Refactor Project::importFromFile to Assets::loadModel
Rewrote the 3D model loading code to be a functional routine outside of the Project class.
I want Project to load .flair files, export .flair files, and copy assets into the project. It should not be in charge of loading new assets. Trying to write .flair logic and 3D model logic into the same functions was getting too confusing.