Join submeshes for objects with multiple materials.
A new mesh will be created and cached for nodes that reference multiple meshes for each material.
Unity displays multiple materials on the same mesh by taking an array of Materials on the MeshRenderer and an array of submeshes on the MeshFilter. Assimp only supplies submeshes so they need to be combined in order for Unity to render everything.