Time an every GameObject loop
Test how long it takes to loop through every GameObject in the scene and check if a teapot should be spawned. Very likely that this is a bad idea, but computers are fast, you never know.
Results: Yes it is indeed terrible. Even without instantiating anything, the loop took 3.94 ms on average which is about half of a 120 fps frame time.