Hook GameObject name change
Test setting a hook GameObject::set_name. It seems like it does catch some name changes, but I can't confirm if it is all of them and it isn't called on instantiation.
If prefabs need to be instantiated on GameObjects with a certain name, they need to be added or removed if the name changes.